Explains and demonstrates prototype inheritance.
Demonstrates how to add custom methods to the Object prototype,
making them accessible to all objects, including arrays and strings.

Also shows how to add custom methods to the Array prototype,
making them only accessible to array objects, not to string and object.

Illustrates the concept of prototype inheritance and the prototype chain
in JavaScript.