Demonstrates call() method usage
Illustrates the use of the `.call()` method to explicitly set the `this` context when invoking a function, allowing properties to be correctly assigned to the desired object.

This ensures that the `username` property is correctly set on the `user` object, resolving the issue of the `username` being undefined in the previous `person` object implementation.