Explains the use of '.call' in Javascript
Clarifies how '.call' allows setting the 'this' context
of a function and passing arguments.

Illustrates how it enables using a function's context
within another function, and the advantages of this.

Explains the use of '.call' in Javascript

Clarifies how '.call' allows setting the 'this' context
of a function and passing arguments.

Illustrates how it enables using a function's context
within another function, and the advantages of this.